Feb 14th Explosion Commit
Added Explosion script scene. Attack Script modified second SetDamage
to GetDamage

+++ Assets/Scripts/Combat/Attack.cs
@@ -93,7 +93,7 @@ public class Attack : MonoBehaviour {
 //		thisDamage = damageMod*damage;
 //	}
 
-	public void SetDamage(int damage) {
+	public void SetDamage(float damage) {
 		thisDamage = damage;
 	}
